Fix `pass/gpu/multiplex_buffers` for loop-invariant buffers
